Move ResourceForm to src/shared
