+++ core-ui/src/shared/ResourceForm/ResourceForm.scss
@@ -11,7 +11,7 @@
   & > .dropdown {
     width: 100%;
   }
-  
+
   .fd-row {
     margin: 0; // override fd's margin: 0 -.25rem;
   }
