+++ core-ui/src/shared/ResourceForm/components/Editor.js
