+++ core-ui/src/shared/ResourceForm/components/ModeSelector.js
